requirements/prod.txt
pytz==2019.1
icalendar==4.0.3
mistune==0.8.4
cython==0.29.9
gunicorn==19.9.0
fastapi==0.26.0
uvicorn==0.7.1
starlette==0.12.0
python-multipart==0.0.5 # for request body form-data parser
email-validator==1.0.4
sentry-asgi==0.2.0

requests-async==0.5.0
uvloop==0.12.2; sys_platform == 'linux'
httptools==0.0.13
gunicorn==19.9.0
peewee-async==0.6.3a0
peewee>=3.9.0,<4.0.0
aiomysql==0.0.20
aioredis==1.2.0


passlib[bcrypt]==1.7.1
pyjwt==1.7.1

# for spider

scrapy-redis
tqdm
scrapy
lxml
python-dateutil

leveldb; sys_platform == 'linux'
